ByteStream refactor
