+++ velox/common/base/CMakeLists.txt
@@ -27,6 +27,8 @@ velox_add_library(
   velox_common_base
   AdmissionController.cpp
   BitUtil.cpp
+  ByteInputStream.cpp
+  ByteOutputStream.cpp
   Counters.cpp
   Fs.cpp
   PeriodicStatsReporter.cpp
