+++ velox/common/memory/tests/ByteStreamTest.cpp
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "velox/common/memory/ByteStream.h"
-
 #include "velox/common/base/tests/GTestUtils.h"
 #include "velox/common/file/FileInputStream.h"
 #include "velox/common/file/FileSystems.h"
+#include "velox/common/memory/BufferInputStream.h"
 #include "velox/common/memory/MmapAllocator.h"
+#include "velox/common/memory/OutputStream.h"
 #include "velox/exec/tests/utils/TempDirectoryPath.h"
 
 #include <gflags/gflags.h>
