+++ velox/common/memory/tests/StreamArenaTest.cpp
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 #include "velox/common/base/tests/GTestUtils.h"
-#include "velox/common/memory/ByteStream.h"
 #include "velox/common/memory/MemoryAllocator.h"
 #include "velox/common/memory/MmapAllocator.h"
+#include "velox/common/memory/OutputStream.h"
 
 #include <gflags/gflags.h>
 #include <gtest/gtest.h>
