+++ velox/serializers/tests/PrestoSerializerTest.cpp
@@ -18,7 +18,6 @@
 #include <gtest/gtest.h>
 #include <vector>
 #include "velox/common/base/tests/GTestUtils.h"
-#include "velox/common/memory/ByteStream.h"
 #include "velox/common/time/Timer.h"
 #include "velox/functions/prestosql/types/TimestampWithTimeZoneType.h"
 #include "velox/vector/fuzzer/VectorFuzzer.h"
